Cleanup and getting
* hooks up test parameter files
* setups up unimplemented interface for DAG implementation
